Move remaining D18 tests to new test file
